bugfix: Wrap shim files in Arc instead of dup'ing
Previously when we forked a process, we'd dup the descriptors kept in
the shadow `Process` object into the new `Process`, and put the new
descriptor numbers in shared memory for the shim-side code of the new
managed process to use.

This is incorrect, though, since the dup'd descriptor doesn't exist in
the new native process.

The original descriptors are transferred between shadow and the managed
process when it's first spawn, using `posix_spawn` actions. It's
simplest to just use those in the new process as well, since the forked
process will also inherit them.