syscall-logger: log blocked syscalls (#3164)
Previously if a syscall blocked, it wasn't recorded in the strace file
until if and when it ran again and completed. Recording blocked syscalls
is slightly redundant since it logs the syscall and parameters multiple
times, but makes it easier to see why control switched to another
thread, and logs syscalls that blocking and never complete, e.g. because
another thread in the process called `exit_group`, the simulation ended,
or in the (unmerged) implementation of execve, which "blocks" and never
returns.

e.g.

```
00:00:01.000000000 [tid 1000] clone(17, 0x0, 0x0, 0x0, 0x0) = 1001
00:00:01.000000000 [tid 1001] write(4, "*", 1) = 1
00:00:01.000000000 [tid 1001] exit_group(...) = <native>
00:00:01.000000000 [tid 1000] read(3, 0x7fffffffe707, 1) = 1
```

->

```
00:00:01.000000000 [tid 1000] clone(17, 0x0, 0x0, 0x0, 0x0) = 1001
00:00:01.000000000 [tid 1000] read(3, 0x7fffffffe707, 1) = <blocked>
00:00:01.000000000 [tid 1001] write(4, "*", 1) = 1
00:00:01.000000000 [tid 1001] exit_group(...) = <native>
00:00:01.000000000 [tid 1000] read(3, 0x7fffffffe707, 1) = 1
```