Test Netlink with `ip addr`
Since `ip addr` uses a netlink socket to get the information of the
interfaces and addresses, we use it to test our Netlink implementation.

We check that the output of the `ip addr` command matches the
ip-addr.stdout file.

We mark the test as an extra because it requirs the iproute2 package.