Moved `core/main.rs` to `shadow.rs`
This is essentially the entrypoint into Shadow, so I think it makes sense to
have this in the top-level directory and not in `core/`. We get a warning if we
name it "main.rs" here since cargo gives that name a special meaning and there
does not seem to be a way to disable that.