alarm syscall: fix handling of Some(0) remaining time
If `alarm` is called at exactly the time when the timer was going to
fire, but before it has fired, we should return 1 instead of 0.

This refactors a bit to make that be true, primarily by returning 0 iff
the timer's `remaining_time` returned `None` (but NOT if it returns
`Some(0s)`).

test