Fix UB in last commit
Getting the two pointers and then writing to them violates the stacked
borrows rules, which would have caused Miri to error if we tested it.

The new code only takes a single pointer at a time, and tests the code
with Miri. This code should be improved in the future.