Increased the callback counter warning limit
Someone ran into these in an actual simulation once. Now that more
objects in Shadow use these `CallbackQueue`s, it's more likely that
there can be a large number of things queued. For example when closing
files in the descriptor table with thousands of sockets, that might
create thousands of callbacks which need to be queued.