When logging errors to stderr, use a different format
I think conventionally stdout should be more of a
standardized/machine-parseable format (as our log records are), and
stderr is intended for human consumption, e.g. on a terminal.

Reformatting the log lines this way when mirroring them to stderr makes
them a little friendlier, and makes it a little easier to understand the
duplication in cases where stdout and stderr are merged.

With an error injected into our sleep test, and stdout redirected to a
file:

```
$ (cd build/src/test/sleep/ && rm -rf sleep-shadow.data       &&  /home/jnewsome/projects/shadow/build/src/main/shadow       --data-directory=sleep-shadow.data       --log-level=trace       --use-cpu-pinning false --strace-logging-mode standard --parallelism 1       /home/jnewsome/projects/shadow/src/test/sleep/sleep.yaml > shadow.log)
** Starting Shadow 3.2.0
Error: process 'testnode.test_sleep.1000' exited with status Normal(101); expected end state was exited: 0 but was exited: 101
Error: Failed to run the simulation
Error:
Error: Caused by:
Error:     1 managed processes in unexpected final state
** Shadow did not complete successfully: Failed to run the simulation
**   1 managed processes in unexpected final state
** See the log for details
```

And in merged output:

```
$ (cd build/src/test/sleep/ && rm -rf sleep-shadow.data       &&  /home/jnewsome/projects/shadow/build/src/main/shadow       --data-directory=sleep-shadow.data       --log-level=trace       --use-cpu-pinning false --strace-logging-mode standard --parallelism 1       /home/jnewsome/projects/shadow/src/test/sleep/sleep.yaml)
...
00:00:00.016603 [507267:shadow-worker] 00:00:01.000000000 [DEBUG] [testnode:11.0.0.1] [process.rs:1486] [shadow_rs::host::process] process 'testnode.test_sleep.1000' has completed or is otherwise no longer running
Error: process 'testnode.test_sleep.1000' exited with status Normal(101); expected end state was exited: 0 but was exited: 101
00:00:00.016660 [507267:shadow-worker] 00:00:01.000000000 [ERROR] [testnode:11.0.0.1] [process.rs:1571] [shadow_rs::host::process] process 'testnode.test_sleep.1000' exited with status Normal(101); expected end state was exited: 0 but was exited: 101
00:00:00.016710 [507267:shadow-worker] 00:00:01.000000000 [TRACE] [testnode:11.0.0.1] [process.rs:750] [shadow_rs::host::process] Not notifying parent of exit: no signal specified
...
```

The duplication in the merged case is still a little weird, but at least
identifiable.