Support Go netlink sockets
When you get the interfaces in Go using `net.Interfaces()`, it uses
netlink sockets to do so.

However, the messages sent by Go sometimes are too short and just ignore
some fields.

For example, the RTM_GETLINK message sent by Go looks like this:
[17, 0, 0, 0, 18, 0, 1, 3, 1, 0, 0, 0, 0, 0, 0, 0, 0]

You can see that `nlmsghdr` took 16 bytes already and it left one byte
at the end as a payload.

In fact, the length should be at least 32 not 17 (16 bytes for
`nlmsghdr` and 16 bytes for `ifinfomsg`).

This PR pads the message with zeroes until it reaches the minimum
length.

I'm not sure how Linux handles it, but I think it's not worth spending
time to investigate more.