Install libglib in workflows that run our rust build scripts
Our rust build scripts look for libglib and fail if it's not present.

It might be possible for them to recognize when they're being run in the
context of these sanitizers and *not* look for libglib, but that seems
like more complexity than it's worth vs just installing libglib.

This unblocks bumping the runner image to 24.04. My guess is that the
22.04 images already had libglib installed for some reason, but the
24.04 ones don't.