Use statically-sized concrete struct for selective acks in C code
This allows us to remove the use of glib's GList, simplifying the
interface between the C TCP stack and the packet. This will make it
easier to replace the C packet with a Rust version. After this, the
entirety of the C TCP header is in a statically-sized struct with
no heap allocated data.