UDP test: start clients after servers
We were starting clients and servers simultaneously, and relying on edge
cases of the shadow scheduler for the server socket to have already been
opened when the client sends a packet.

This breaks if we use `--model-unblocked-syscall-latency`, since the
timings are perturbed.

If we really want to exercise this edge case, it'd probably be best to
use a dedicated test with `--model-unblocked-syscall-latency` explicitly
disabled.

Rearranges such that processes appear in the order that they're started,
and that none are started simultaneously.