shim: Track execution context more precisely
Previously we deferred setting the execution context to Shadow (enabling
native syscalls) until just before we know we need to make native
syscalls.

Now that we're more broadly trying to track when we're executing within
the context of shadow code, it makes more sense to push this out to the
"entry points" of shadow shim code.

These entry points include:

* shim signal handlers
* public API functions (such as those called from the libc preload and
  injector libraries)

This also adds `debug_assert`s in some internal functions to validate
that we're in the shadow execution context as expected, especially those
manipulating global state. This is to ensure that signal handlers can
correctly identify whether the "calling" context is shadow code or
managed code.