Move shadow syscall numbers into shadow-shim-helper-rs
Putting it in this crate is more sensible in terms of dependencies,
since we also use these from the shim.

This also moves the definitions from C to Rust.

Since the type-name is used when referencing these in Rust, and prefixed
into the C bindings, the individual enumerator names can be a little
shorter.

We keep the `shadow_` in `shadow_yield` though, because the `yield`
token is reserved and even if it weren't (or if we dealt with escaping
it) might be otherwise confusing.