Fix the issue where loopback-bound sockets attempt to connect to a public IP (#3531)
In the current sys_connect handler, sockets bound to localhost are not
handled correctly.

https://github.com/shadow/shadow/blob/3e478a9cd1b3277650a6d957a22ddb8de87e6e34/src/main/host/descriptor/socket/inet/legacy_tcp.rs#L760-L765

Shadow does not check cases where it connects to an external address
using a socket bound to 127.0.0.1, which later causes crashes when the
worker fails to resolve the localhost address to a host.

https://github.com/shadow/shadow/blob/3e478a9cd1b3277650a6d957a22ddb8de87e6e34/src/main/core/worker.rs#L356-L358

The same issue also occurs with UDP (and possibly in tcp.rs).

This PR brings a fix to this issue, although I think we'd want to mimic
a routing table and perform these handling there.