Make shadow-shim-helper-rs no_std by default
We want this to be `no_std` in the shim. We currently do compile with
`std` for the main shadow crate, only so that in the main shadow
process we can detect if the current thread is already panicking in
`Drop<rooted_cell::rc::RootedCell>` and avoid extra noisy panic
messages.

Progress on https://github.com/shadow/shadow/issues/2919