getauxval test: a few robustness tweaks
I had reverted skipping testing `AT_HWCAP2` since it seems to return the
same value via glibc as /proc, but looking at the glibc code this might
not always hold. Probably better to skip trying to compare it after all.

Also test more precisely whether we expect libc to find a value of 0 vs
not find a value at all (in which case it still returns 0 but also sets
errno to ENOENT).