Removed use of `SOCK_PACKET` in tests
This is deprecated in the libc crate now, and has been "not recommended"
in linux for many years. Just because it's deprecated from libc doesn't
mean that we shouldn't test the kernel API, but this socket type is not
supported by Shadow anyways and hopefully no applications are using it.