DescriptorTable::remove_range: return descriptors in deterministic order
Previously we were returning them in the `HashMap`'s arbitrary order,
and then in `SyscallHandler::close_range` closing descriptors and
pushing callbacks in that arbitrary order.

Changing the data structure to `BTreeMap` ensures we process the
descriptors in a deterministic (sorted) order. Alternatively we could
just sort the vector returned by `DescriptorTable::remove_range`, but
using a sorted data structure in the first place will make it harder for
future subtle bugs of this nature to sneak back in.

This seems to fix the nondeterministic scheduling (presumably due to
callbacks being called in different order) and nondeterministic file
descriptor assignments observed in #3610.