Add some /proc handling (#3613)
Remap /proc/self/* paths to point to the correct native path. Notably
this fixes behavior for reading `/proc/self/maps`, which the rust
runtime does, and for reading `/proc/self/exe` which asan does.

Log a warning when opening proc files that we haven't specifically
tested, thought about, or otherwise vetted.

Progress on #3541
Fixes #3536
Progress on #2790