TCP: Un-parent child sockets once they've been accepted
Once a child socket has been `accept`ed (associated with a file
descriptor and moved off of the listening socket's queue), we now remove
the parent-child relationship and associate the local:remote address
pair directly with the corresponding network interface, instead of
continuing to route packets to the listening socket and having the
listening socket route them to the child.

This allows a listening socket to be completely cleaned up when it's
closed instead of keeping it around to route packets to the child
sockets, which allows us to disassociate and subsequently reuse the
listening address.

Fixes https://github.com/shadow/shadow/issues/3563