TCP: Un-parent child sockets once they've been accepted
Once a child socket has been `accept`ed (associated with a file
descriptor and moved off of the listening socket's queue), we now remove
the parent-child relationship and associate the local:remote address
pair directly with the corresponding network interface, instead of
continuing to route packets to the listening socket and having the
listening socket route them to the child.

This allows a listening socket to be completely cleaned up when it's
closed instead of keeping it around to route packets to the child
sockets, which allows us to disassociate and subsequently reuse the
listening address.

This is progress on https://github.com/shadow/shadow/issues/3563, though
I think there are still a couple related issues:

* The test still requires some non-zero delay somewhere between
  accepting a connection and trying to reuse the listening socket's
  address. This probably won't matter for real use-cases, but I'm not
  sure why it's the case.
* We (still) don't close or reset connections that are in the
  listen-queue when the listening socket has closed. I think it's
  somewhat better now in that if the client end of the connection sends
  any packets, they should now get a RST since there's no corresponding
  entry in the routing table. But if the client is waiting to receive data
  it might wait indefinitely (which I think was also the case
  previously).