syscall wrappers: don't wrap getrlimit or setrlimit
The libc implementations of these functions actually use different
struct definitions than the kernel.

On some platforms, the kernel definition only has 32-bit values.

On modern x86_64 kernels, the kernel definition is equivalent to the
64-bit values, so the old behavior of wrapping these was *probably* ok.

glibc actually implements both of these functions using the prlimit64
syscall whether or not it's on a platform with the 32/64 bit
discrepancy, though the ultimate behavior should be the same on
platforms without that discrepancy.

All this to say, skipping the wrapper here probably won't actually
change any simulation behavior, but seems prudent given all the caveats
and subtleties.